% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packs.R
\name{chrom_state_grange}
\alias{chrom_state_grange}
\alias{chrome_state_grange}
\title{Chromatin states manipulation}
\usage{
chrom_state_grange(csvFilePath)
}
\arguments{
\item{csvFilepath}{path of csv file}
}
\description{
return grange
}
