% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packs.R
\name{gwas_grange_from_file}
\alias{gwas_grange_from_file}
\title{Gwas catalog manipulation}
\usage{
gwas_grange_from_file(file, gap)
}
\arguments{
\item{file}{csv file}

\item{gap}{the gap bps across the position of SNP}
}
\description{
return grange
}
