% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packs.R
\name{gwas_grange_from_path}
\alias{gwas_grange_from_path}
\alias{gwas_grange}
\title{Gwas catalog manipulation}
\usage{
gwas_grange_from_path(csvFilePath, gap)
}
\arguments{
\item{gap}{the gap bps across the position of SNP}

\item{csvFilepath}{path of csv file}
}
\description{
return grange
}
